+++ CHANGELOG.md
@@ -5,7 +5,9 @@
  - Added priority to map urls, default is to use mcserver throttling
  - Added visual effects to absorb spell
  - Add permission check to arrow spell
-
+ - Add some work-arounds for players using the armor slots to smuggle items in/out of the wand inventory.
+ - Add a work-around for a player placing a block out of their wand inventory.
+ 
 ## 2.8.3
 
  - Add /wand combine command